pyproject.toml
[tool.poetry]
name = "nautobot-plugin-chatops-ansible"
version = "0.0.1"
description = "Nautobot Chatpops Ansible Tower integration"
authors = ["Network to Code, LLC <info@networktocode.com>"]

readme = "README.md"
homepage = "https://github.com/networktocode-llc/nautobot-plugin-chatops-ansible"
repository = "https://github.com/networktocode-llc/nautobot-plugin-chatops-ansible"
keywords = ["nautobot", "nautobot-plugin"]
include = [
    "README.md",
]
packages = [
    { include = "nautobot_plugin_chatops_ansible" },
]

[tool.poetry.plugins."nautobot.workers"]
"ansible" = "nautobot_plugin_chatops_ansible.worker:ansible"

[tool.poetry.dependencies]
python = "^3.6"
invoke = "^1.4.1"
nautobot-chatops = "^1.1.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
yamllint = "^1.23.0"
bandit = "^1.7.0"
pylint = "^2.5.2"
pylint-django = "^2.0.15"
pydocstyle = "^5.0.2"

[tool.black]
line-length = 120
target-version = ['py36']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | settings.py     # This is where you define files that should not be stylized by black
                     # the root of the project
)
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

